Add files via upload
Recommend five articles with high popularity on the homepage.We added the homepage into two parts. Articles in the first part are top 5 in popularity, and this part will update according to the popularity of each article and how long the article has been shown on it Articles in the second part are all the articles posted before,sorted by release time from newest to oldest.